Implemented & Structed Dashboard Component

+++ DTOs/TaskDTOs/ClaimHistoryDto.cs
@@ -3,7 +3,7 @@
     public class ClaimHistoryDto
     {
         public Guid Id { get; set; }
-        public int PointsClaimed { get; set; }
+        public decimal PointsClaimed { get; set; }
         public decimal ConvertedAmount { get; set; }
         public DateTime ClaimedAt { get; set; }
     }
