+++ DTOs/TaskDTOs/TaskItemDto.cs
@@ -5,7 +5,7 @@
         public Guid Id { get; init; }
         public string Title { get; init; } = string.Empty;
         public string Description { get; init; } = string.Empty;
-        public int Points { get; init; }
+        public decimal Points { get; init; }
         public bool IsActive { get; init; }
     }
 }
